+++ modules/statsite/variables.tf
@@ -1,0 +1,13 @@
+variable "key_name" {}
+variable "platform_instance_id" {}
+variable "proxy_endpoint" {}
+variable "statsite_amis" {}
+variable "statsite_instance_type" {}
+variable "statsite_subnet_id_1" {}
+variable "statsite_user_init" {}
+variable "statsite_volume_az" {}
+variable "statsite_volume_force_detach" {}
+variable "statsite_volume_skip_destroy" {}
+variable "statsite_volume_size" {}
+variable "statsite_volume_snap" {}
+variable "ssh_cidr_blocks" {}